// server/src/modules/users/user.enums.ts
export enum UserRole {
  STUDENT = "student",
  INSTRUCTOR = "instructor",
  INSTITUTE_ADMIN = "institute_admin",
  SUPER_ADMIN = "super_admin",
}

export enum UserStatus {
  ACTIVE = "active",
  INACTIVE = "inactive",
  SUSPENDED = "suspended",
  DELETED = "deleted",
}

export enum Gender {
  MALE = "male",
  FEMALE = "female",
  OTHER = "other",
  PREFER_NOT_TO_SAY = "prefer_not_to_say",
}